@@ -9,7 +9,7 @@ import { RegisterComponent } from './component/register/register.component';
 import { WelcomePageComponent } from './component/welcome-page/welcome-page.component';
 import { FormsModule,ReactiveFormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { AdminComponent } from './component/admin/admin.component';
 import { UserComponent } from './component/user/user.component';
 import { LoaderComponent } from './component/loader/loader.component';
@@ -20,31 +20,24 @@ import { ForgetPasswordComponent } from './component/forget-password/forget-pass
 import { TicketPopupComponent } from './component/ticket-popup/ticket-popup.component';
 import { NgbModalModule } from '@ng-bootstrap/ng-bootstrap';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    LoginComponent,
-    HomeComponent,
-    RegisterComponent,
-    WelcomePageComponent,
-    AdminComponent,
-    UserComponent,
-    LoaderComponent,
-    TicketComponent,
-    ForgetPasswordComponent,
-    TicketPopupComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    ReactiveFormsModule,
-    HttpClientModule ,
-    BrowserAnimationsModule,
-    FormsModule,
-    FontAwesomeModule,
-    NgbModalModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        LoginComponent,
+        HomeComponent,
+        RegisterComponent,
+        WelcomePageComponent,
+        AdminComponent,
+        UserComponent,
+        LoaderComponent,
+        TicketComponent,
+        ForgetPasswordComponent,
+        TicketPopupComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        ReactiveFormsModule,
+        BrowserAnimationsModule,
+        FormsModule,
+        FontAwesomeModule,
+        NgbModalModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
